chore: Moved rate_limit_exceeded_callback() to src/extensions.py from src/Routes/routes_fastapi.py
